+++ crates/sncast/src/main.rs
@@ -39,11 +39,10 @@ use sncast::response::transformed_call::transform_response;
 use sncast::response::ui::UI;
 use sncast::{
     ValidatedWaitParams, WaitForTx, get_account, get_block_id, get_class_hash_by_address,
-    get_contract_class,
+    get_contract_class, with_account,
 };
 use starknet_commands::ledger::{self, Ledger};
 use starknet_commands::verify::Verify;
-use starknet_rust::accounts::Account;
 use starknet_rust::core::types::ContractClass;
 use starknet_rust::core::types::contract::{AbiEntry, SierraClass};
 use starknet_rust::core::utils::get_selector_from_name;
@@ -264,6 +263,7 @@ fn init_logging() {
         .with_filter(
             EnvFilter::builder()
                 .with_default_directive(LevelFilter::WARN.into())
+                .with_default_directive("coins_ledger=off".parse().expect("valid directive"))
                 .with_env_var("SNCAST_LOG")
                 .from_env_lossy(),
         );
@@ -307,14 +307,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
 
             let rpc = declare.common.rpc.clone();
 
-            let account = get_account(
-                &config,
-                &provider,
-                &declare.common.rpc,
-                config.keystore.as_ref(),
-                ui,
-            )
-            .await?;
+            let account = get_account(&config, &provider, &declare.common.rpc, ui).await?;
             let manifest_path = assert_manifest_path_exists()?;
             let package_metadata = get_package_metadata(&manifest_path, &declare.package)?;
             let artifacts = build_and_load_artifacts(
@@ -330,17 +323,19 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
             )
             .expect("Failed to build contract");
 
-            let result = starknet_commands::declare::declare(
-                declare.contract_name.clone(),
-                declare.common.fee_args,
-                declare.common.nonce,
-                &account,
-                &artifacts,
-                wait_config,
-                false,
-                ui,
-            )
-            .await
+            let result = with_account!(&account, |account| {
+                starknet_commands::declare::declare(
+                    declare.contract_name.clone(),
+                    declare.common.fee_args,
+                    declare.common.nonce,
+                    account,
+                    &artifacts,
+                    wait_config,
+                    false,
+                    ui,
+                )
+                .await
+            })
             .map_err(handle_starknet_command_error)
             .map(|result| match result {
                 DeclareResponse::Success(declare_transaction_response) => {
@@ -402,24 +397,19 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
                 }
             };
 
-            let account = get_account(
-                &config,
-                &provider,
-                &declare_from.common.rpc,
-                config.keystore.as_ref(),
-                ui,
-            )
-            .await?;
+            let account = get_account(&config, &provider, &declare_from.common.rpc, ui).await?;
 
-            let result = starknet_commands::declare_from::declare_from(
-                contract_source,
-                &declare_from.common,
-                &account,
-                wait_config,
-                false,
-                ui,
-            )
-            .await
+            let result = with_account!(&account, |account| {
+                starknet_commands::declare_from::declare_from(
+                    contract_source,
+                    &declare_from.common,
+                    account,
+                    wait_config,
+                    false,
+                    ui,
+                )
+                .await
+            })
             .map_err(handle_starknet_command_error)
             .map(|result| match result {
                 DeclareResponse::Success(declare_transaction_response) => {
@@ -455,8 +445,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
 
             let provider = rpc.get_provider(&config, ui).await?;
 
-            let account =
-                get_account(&config, &provider, &rpc, config.keystore.as_ref(), ui).await?;
+            let account = get_account(&config, &provider, &rpc, ui).await?;
 
             let (class_hash, declare_response) = if let Some(class_hash) = identifier.class_hash {
                 (class_hash, None)
@@ -476,22 +465,23 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
                 )
                 .expect("Failed to build contract");
 
-                let declare_result = declare(
-                    contract_name,
-                    fee_args.clone(),
-                    nonce,
-                    &account,
-                    &artifacts,
-                    WaitForTx {
-                        wait: true,
-                        wait_params: wait_config.wait_params,
-                        // Only show outputs if user explicitly provides `--wait` flag
-                        show_ui_outputs: wait_config.wait,
-                    },
-                    true,
-                    ui,
-                )
-                .await
+                let declare_result = with_account!(&account, |account| {
+                    declare(
+                        contract_name,
+                        fee_args.clone(),
+                        nonce,
+                        account,
+                        &artifacts,
+                        WaitForTx {
+                            wait: true,
+                            wait_params: wait_config.wait_params,
+                            show_ui_outputs: wait_config.wait,
+                        },
+                        true,
+                        ui,
+                    )
+                    .await
+                })
                 .map_err(handle_starknet_command_error);
 
                 // Increment nonce after successful declare if it was explicitly provided
@@ -529,18 +519,20 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
             let arguments: Arguments = arguments.into();
             let calldata = arguments.try_into_calldata(&contract_class, &selector)?;
 
-            let result = starknet_commands::deploy::deploy(
-                class_hash,
-                &calldata,
-                salt,
-                unique,
-                fee_args,
-                nonce,
-                &account,
-                wait_config,
-                ui,
-            )
-            .await
+            let result = with_account!(&account, |account| {
+                starknet_commands::deploy::deploy(
+                    class_hash,
+                    &calldata,
+                    salt,
+                    unique,
+                    fee_args,
+                    nonce,
+                    account,
+                    wait_config,
+                    ui,
+                )
+                .await
+            })
             .map_err(handle_starknet_command_error);
 
             let result = if let Some(declare_response) = declare_response {
@@ -616,8 +608,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
 
             let provider = rpc.get_provider(&config, ui).await?;
 
-            let account =
-                get_account(&config, &provider, &rpc, config.keystore.as_ref(), ui).await?;
+            let account = get_account(&config, &provider, &rpc, ui).await?;
 
             let selector = get_selector_from_name(&function)
                 .context("Failed to convert entry point selector to FieldElement")?;
@@ -627,17 +618,19 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
 
             let calldata = arguments.try_into_calldata(&contract_class, &selector)?;
 
-            let result = starknet_commands::invoke::invoke(
-                contract_address,
-                calldata,
-                nonce,
-                fee_args,
-                selector,
-                &account,
-                wait_config,
-                ui,
-            )
-            .await
+            let result = with_account!(&account, |account| {
+                starknet_commands::invoke::invoke(
+                    contract_address,
+                    calldata,
+                    nonce,
+                    fee_args,
+                    selector,
+                    account,
+                    wait_config,
+                    ui,
+                )
+                .await
+            })
             .map_err(handle_starknet_command_error);
 
             let block_explorer_link =
@@ -728,14 +721,7 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
 
         Commands::Balance(balance) => {
             let provider = balance.rpc.get_provider(&config, ui).await?;
-            let account = get_account(
-                &config,
-                &provider,
-                &balance.rpc,
-                config.keystore.as_ref(),
-                ui,
-            )
-            .await?;
+            let account = get_account(&config, &provider, &balance.rpc, ui).await?;
 
             let result =
                 starknet_commands::balance::balance(account.address(), &provider, &balance).await?;